Re-organise Model tests
